Sync messages sent from other devices
When signal-cli syncs messages that have been sent from the user's other
devices, an envelope is produced with the message at
envelope.syncMessage.sentMessage.
These are a form of sync messages, whose information is provided under
envelope.syncMessage, as compared to normal messages which are under
envelope.dataMessage.

Messages are now inserted into the chat history depending on their
timestamp, ensuring that they appear chronologically.
The timestamp of sent messages has been adjusted to be in milliseconds
instead of seconds, as timestamps from signal-cli are in milliseconds.

Accesses to envelope.dataMessage have been adjusted to also check
envelope.syncMessage.sentMessage so the client can properly handle
synced messages. This is somewhat messy and could be improved in the
future with a mechanism for extracting a 'message' from an envelope,
which would check for both dataMessage and syncMessage.sentMessage.