Send messages asynchronously in a background process
Makes the sending subprocess non-blocking, so that scli remains
interactive while signal-cli handles sending the message in the
background. Several outgoing messages can be processed concurrently.

To get the response from the sending process, urwid's
`MainLoop.watch_pipe()` is used to call the handler function with the
subprocess's return code when it exits. An alternative approach would
be to register the `SIGCHLD` signal handler using the `signal` module
from the standard library. With using either the `watch_pipe` or the
`SIGCHLD` handler, one of the `os.wait()` family of functions can be
used to get the exit code; or the container of all the sending process
can be iterated over and the status of the processes checked. However,
the approach taken here seems to be the cleanest of all the
alternatives.