Create AsyncProc and Daemon classes
`AsyncProc.run()` is a general method for executing a subprocess in the background shell and running a callback function on the subprocess exit.
`Daemon` is a specialization of AsyncProc that handles interactions with signal-cli's daemon.

Most of the changes is moving the existing code around, and generalizing it for any kind of background processes and callback functions.