Determine when signal-cli daemon has started from daemon's stderr
As of v0.7.4, the signal-cli daemon writes a string to stderr once the dbus service has started. Use it instead of periodically polling a list of active dbus services to check when the daemon has started.

Still might need to use polling *after* the daemon prints the line to stderr, because sometimes at the time that line is written, the dbus service is not actually registered yet.