Fix message status for groups with unregistered users
So if some of the group members deletes their signal account after
getting into a group, they stay in the group but they will never be
able to receive a message, hence dbus call will exit with a non-zero
status code. So this simply checks if failed send operation is due to
unregisterd users and if so instead of marking it failed, it lets the
message through. Problem is message will never be marked as read
because of those unregistered users. I haven't been able to come up
with a clear solution for this, becuase the group member receipt seems
logic seems quite complicated at this point. But this is better than
nothing, at least seeing one tick rather than an ugly X is nicer.