Add LazyEvalListWalker class
This urwid.ListWalker subclass allows to defer the creation of widgets until they are in ListBox's view. With it, the widgets (e.g. message widgets in a chat) do not have to be created until they can be seen.
This can speed up program's launch: the message widgets from history don't have to be initialized at startup.